Added Firefox support
auth.py now looks for ~/.mozilla/firefox/ and if that file/directory
exists then it will use pycookiecheat's firefox function instead. It
then checks for ~/.config/google-chrome and then run's the original
chrome cookie function and if that fails then it will attempt to find
the cookie from a local file.

This is totally a hack specifically to fix my problem and I'm new to
contributing to open source so please take this with a huge grain of
salt :)

Someone else may have also added FF support but I didn't notice...